+++ robotidy/transformers/ReplaceRunKeywordIf.py
@@ -87,7 +87,7 @@ class ReplaceRunKeywordIf(ModelTransformer):
             Token(Token.EOL)
         ])
         prev_if = None
-        for branch in reversed(list(self.split_args_on_delimeters(raw_args, ('ELSE', 'ELSE IF')))):
+        for branch in reversed(list(self.split_args_on_delimeters(raw_args, ('ELSE', 'ELSE IF'), assign=assign))):
             if branch[0].value == 'ELSE':
                 header = ElseHeader([
                     separator,
@@ -128,7 +128,7 @@ class ReplaceRunKeywordIf(ModelTransformer):
     def create_keywords(self, arg_tokens, assign, indent):
         if normalize_name(arg_tokens[0].value) == 'runkeywords':
             return [self.args_to_keyword(keyword[1:], assign, indent)
-                    for keyword in self.split_args_on_delimeters(arg_tokens, ('AND',))]
+                    for keyword in self.split_args_on_delimeters(arg_tokens, ('AND',), [])]
         return self.args_to_keyword(arg_tokens, assign, indent)
 
     def args_to_keyword(self, arg_tokens, assign, indent):
@@ -140,10 +140,14 @@ class ReplaceRunKeywordIf(ModelTransformer):
         return KeywordCall.from_tokens(separated_tokens)
 
     @staticmethod
-    def split_args_on_delimeters(args, delimeters):
+    def split_args_on_delimeters(args, delimeters, assign):
         split_points = [index for index, arg in enumerate(args) if arg.value in delimeters]
         prev_index = 0
         for split_point in split_points:
             yield args[prev_index:split_point]
             prev_index = split_point
         yield args[prev_index:len(args)]
+        if 'ELSE' in delimeters and assign:
+            if not any(arg.value == 'ELSE' for arg in args):
+                values = [Token(Token.ARGUMENT, '${None}')] * len(assign)
+                yield [Token(Token.ELSE), Token(Token.ARGUMENT, 'Set Variable'), *values]